pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "dotenv-linter==0.5.0",
    "load-dotenv==0.1.0",
]
